Cart issues fixed

+++ app/Http/Controllers/Api/V1/CartController.php
@@ -46,7 +46,11 @@ class CartController extends Controller
     {
         $validator = Validator::make($request->all(), [
             'guest_id' => $request->user ? 'nullable' : 'required',
-            'item_id' => 'required|integer',
+            'item_id' => [
+              'required',
+               Rule::exists('food', 'id')->whereNull('deleted_at'),
+             ],
+
             'model' => 'required|string|in:Food,ItemCampaign',
             'price' => 'required|numeric',
             'variation_options' => 'nullable|array',
@@ -54,11 +58,17 @@ class CartController extends Controller
             'restaurant_id' => [
                   'required',
                    Rule::exists('restaurants', 'id')->whereNull('deleted_at'),
-                ]
+            ]
         ]);
 
         if ($validator->fails()) {
             return response()->json(['status'=>'failed','errors' => Helpers::error_processor($validator)], 403);
+        }
+
+        $foodData=Food::where('id', $request->item_id)->where('restaurant_id', $request->restaurant_id)->first();
+
+        if(is_null($foodData)){
+            return response()->json(['status'=>'failed','message'=>"Food details not found for selected restaurant"], 400);  
         }
 
         $user_id = $request->user ? $request->user->id : $request['guest_id'];
@@ -127,9 +137,15 @@ class CartController extends Controller
             return response()->json(['errors' => Helpers::error_processor($validator)], 403);
         }
 
+
         $user_id = $request->user ? $request->user->id : $request['guest_id'];
         $is_guest = $request->user ? 0 : 1;
-        $cart = Cart::find($request->cart_id);
+        $cart = Cart::where('id', $request->cart_id)->first();
+  
+        if(is_null($cart)){
+            return response()->json(['status'=>'failed','message'=>"Cart details not found"], 400);  
+        }
+
         $item = $cart->item_type === 'App\Models\Food' ? Food::find($cart->item_id) : ItemCampaign::find($cart->item_id);
         if($item->maximum_cart_quantity && ($request->quantity>$item->maximum_cart_quantity)){
             return response()->json(['status'=>'failed', 'code' => 'cart_item_limit', 'message' => translate('messages.maximum_cart_quantity_exceeded')], 403);
@@ -182,7 +198,11 @@ class CartController extends Controller
         $user_id = $request->user ? $request->user->id : $request['guest_id'];
         $is_guest = $request->user ? 0 : 1;
 
-        $cart = Cart::find($request->cart_id);
+        $cart = Cart::where('id', $request->cart_id)->first();
+        if(is_null($cart)){
+            return response()->json(['status'=>'failed','message'=>"Cart details not found"], 400);  
+        }
+
         $cart->delete();
 
         $carts = Cart::where('user_id', $user_id)->where('is_guest',$is_guest)->get()
